Remove state from BaseRadioButtonGroup
