+++ tests/RadioButtonGroup/BaseRadioButtonGroup.test.tsx
@@ -153,57 +153,4 @@ describe('<BaseRadioButtonGroup/>', () => {
       expect(onChange).toHaveBeenCalledWith('b')
     })
   })
-
-  describe('When a user interacts with a <BaseRadioButtonGroup>', () => {
-    test('When unchecked <RadioButton> is clicked Then it becomes checked', () => {
-      const onChange = jest.fn()
-
-      const { getByRole } = render(
-        <BaseRadioButtonGroup onChange={onChange}>
-          <RadioButton disabled={false} value="a" />
-        </BaseRadioButtonGroup>
-      )
-
-      const radioButton = getByRole('radio')
-      fireEvent.click(radioButton)
-
-      expect(radioButton).toHaveAttribute('aria-checked', 'true')
-    })
-
-    test('When <RadioButtons> are toggled Then they transition consistently', () => {
-      let checked = 'a'
-      const onChange = (value: string) => {
-        checked = value
-      }
-
-      render(
-        <BaseRadioButtonGroup checked={checked} onChange={onChange}>
-          <div id="a">
-            <RadioButton disabled={false} value="a" />
-          </div>
-          <div id="b">
-            <RadioButton disabled={false} value="b" />
-          </div>
-        </BaseRadioButtonGroup>
-      )
-      const radioButtonA = document.getElementById('a').firstChild
-      const radioButtonB = document.getElementById('b').firstChild
-
-      expect(checked).toBe('a')
-      expect(radioButtonA).toHaveAttribute('aria-checked', 'true')
-      expect(radioButtonB).not.toHaveAttribute('aria-checked', 'true')
-
-      fireEvent.click(radioButtonB)
-
-      expect(checked).toBe('b')
-      expect(radioButtonA).not.toHaveAttribute('aria-checked', 'true')
-      expect(radioButtonB).toHaveAttribute('aria-checked', 'true')
-
-      fireEvent.click(radioButtonA)
-
-      expect(checked).toBe('a')
-      expect(radioButtonA).toHaveAttribute('aria-checked', 'true')
-      expect(radioButtonB).not.toHaveAttribute('aria-checked', 'true')
-    })
-  })
 })
